Version 1.1.5, new landing page, other improvements
Added dummy login/logout functionality for testing. New landing page with some key dummy insights. Made it so that themes persist on page reload. Added a new component CustomTable with the option to dynamically size it using a ref. Modified navbar so it only shows when "logged in". Added protected routes and error boundary which displays ISE page. Kind of all for now.