UI refinements
Upgraded the expand item functionality, added a new size to charts (xl) for modals, enlarged the mainblock size from 5xl to 6xl width. Standardised buttons across the site, fixed the reset form button to only be available on form entry (even if incomplete). Increased the size of the navbar, added error logging to the 500 code page.