Enforcing DRY, UserRoleProtectedRoute, UI/UX
Added Zustand to the portfolio and analysis stuff to mock fetch portfolio details to render. Added UserRoleProtectedRoute for checking if a user has a certain role in App.jsx. Modified Navbar in some sort of way. Added a class to CustomTable which prevents a scrollbar from being rendered as it was really annoying me. Added more mock log in stuff such as setting the fullname, username, email, etc., of the user on log in for use in Profile Card. Some solid work today tbf. Probably some other stuff too but i forgot, all looks good and mostly works for now.