Integrated ticker search in trade booking card!
One of the coolest things ive got so far. Quick commit rundown: created a useassetdatahook (not gonna use for now), created AssetDataService for fetching asset metadata (for now). FlagMapper for mapping country codes to tickers (as tiingo gives countryCode instead of exchange, slightly annoying but it will have to do). Modified tradehistorycard to have white text on SELL trades. Created a FuzzyAssetMetadataSearch bar for searching for assets and displaying a dropdown with asset metadata. That's been integrated into BookTradeCard. New CardOneTooltip (cardone with a tooltip). I think thats it for now. Thanks