Checkpoint: cash service, next is orchestrator
This commit is merely a checkpoint of sorts. I've created a cash flow model (cash_ledger table), and created a cash_service. That all works individually, but since the system doesn't know if a trade is going to open a new position, close, update, overclose, etc., we need an orchestrator which will tie it all together.