Optional fx+conv. in cash_ledger, autodeposit on portfolio setup
First i will say that i deleted a bunch of duplicated files. Im not sure how they got here but it wasn't intentional. Secondly, I made fx rate and converted amount optional in the cash ledger table. We don't need those kind of things on deposits and withdraws, and it's just clutter having artbitrary values (e.g., who knew that a deposit of USD100 has an FX rate of 1.00 and a converted amount of USD100?!!). Also integrated _record_initial_deposit into the portfolio system, now when you update the portfolio, it checks if initial_cash is present, and if so it passes it to the CashService service. Thanks