Free cash (excludes short encumbered) + tests
Added root_portfolio_id to position and cash_ledger tables. This is only done for partial closes, as it allows the system to trace whats actually happening when a synthetic position from a partial close is created. This ties into the next thing: cash_service

root_portfolio_id in positions and cash_ledger, short encumberment

Added root_portfolio_id in the positions and cash ledger. This is to enable tracking of synthetic positions which are created as a result of partial closes. This is great for the future, and also critical for the next milestone: free cash. Free cash is now calculated as net balance - net encumberement from short proceeds. The problem was that sell trades added cash (by design, this is trivial), but that cash should not be allowed to be spent on creating new positions (as it is margin). The solution is encumberement. To implement encumberement, the cash service finds every root short position (original position/parent position) which is still open, and for those roots it sums all short sell cash inflows. Then it subtracts all short cover cash flows already used to buy back part or all of those shorts. Then it sums the remaining amount acorss all open shorts (as SHORT COVER uses a negative sign), and the result is how much short-sale cash is not actually free to reuse. I added some tests (including a relatively complex one, namely test_encumbered_multishort_complex.py) and the results are phenomenal - it works. Also modified the usershowallcard to prevent the null is not defined bug when the user is not logged in, createportfolio card for initialising a portfolio without a description. Thats all for now, thanks